Move `MockAuthProvider` to `testbase.http`
There's nothing specific to the auth extension in the mock HTTP server
implementation and it will be useful in tests of other HTTP extensions,
so move it to `testbase.http` and rename to `MockHttpServer`.
